refactor(lsp): Add state to lsp
Move lsp config from defaults.lua to lsp.config
