+++ lua/lsp/init.lua
@@ -1,20 +1,27 @@
-local M = {}
+local M = {
+  config = {},
+}
+
 local Log = require "core.log"
 
-function M.setup(config)
-  vim.lsp.protocol.CompletionItemKind = config.completion.item_kind
+function M:setup(overrides)
+  local lsp_config = require "lsp.config"
+  local Config = require "config"
+  self.config = Config(lsp_config):merge(overrides).entries
 
-  for _, sign in ipairs(config.diagnostics.signs.values) do
+  vim.lsp.protocol.CompletionItemKind = self.config.completion.item_kind
+
+  for _, sign in ipairs(self.config.diagnostics.signs.values) do
     vim.fn.sign_define(sign.name, { texthl = sign.name, text = sign.text, numhl = sign.name })
   end
 
-  require("lsp.handlers").setup()
+  require("lsp.handlers").setup {
+    diagnostics = self.config.diagnostics,
+    popup_border = self.config.popup_border,
+  }
 end
 
 local function lsp_highlight_document(client)
-  if lvim.lsp.document_highlight == false then
-    return -- we don't need further
-  end
   -- Set autocommands conditional on server_capabilities
   if client.resolved_capabilities.document_highlight then
     vim.api.nvim_exec(
@@ -34,28 +41,30 @@ local function lsp_highlight_document(client)
 end
 
 local function add_lsp_buffer_keybindings(bufnr)
-  local status_ok, wk = pcall(require, "which-key")
-  if not status_ok then
-    return
-  end
+  local which_key = require "core.builtins.which-key"
 
-  local keys = {
-    ["K"] = { "<cmd>lua vim.lsp.buf.hover()<CR>", "Show hover" },
-    ["gd"] = { "<cmd>lua vim.lsp.buf.definition()<CR>", "Goto Definition" },
-    ["gD"] = { "<cmd>lua vim.lsp.buf.declaration()<CR>", "Goto declaration" },
-    ["gr"] = { "<cmd>lua vim.lsp.buf.references()<CR>", "Goto references" },
-    ["gI"] = { "<cmd>lua vim.lsp.buf.implementation()<CR>", "Goto Implementation" },
-    ["gs"] = { "<cmd>lua vim.lsp.buf.signature_help()<CR>", "show signature help" },
-    ["gp"] = { "<cmd>lua require'lsp.peek'.Peek('definition')<CR>", "Peek definition" },
-    ["gl"] = {
-      "<cmd>lua require'lsp.handlers'.show_line_diagnostics()<CR>",
-      "Show line diagnostics",
+  which_key:register({
+    normal_mode = {
+      ["K"] = { "<cmd>lua vim.lsp.buf.hover()<CR>", "Show hover" },
+      ["gd"] = { "<cmd>lua vim.lsp.buf.definition()<CR>", "Goto Definition" },
+      ["gD"] = { "<cmd>lua vim.lsp.buf.declaration()<CR>", "Goto declaration" },
+      ["gr"] = { "<cmd>lua vim.lsp.buf.references()<CR>", "Goto references" },
+      ["gI"] = { "<cmd>lua vim.lsp.buf.implementation()<CR>", "Goto Implementation" },
+      ["gs"] = { "<cmd>lua vim.lsp.buf.signature_help()<CR>", "show signature help" },
+      ["gp"] = { "<cmd>lua require'lsp.peek'.Peek('definition')<CR>", "Peek definition" },
+      ["gl"] = {
+        "<cmd>lua require'lsp.handlers'.show_line_diagnostics()<CR>",
+        "Show line diagnostics",
+      },
     },
-  }
-  wk.register(keys, { mode = "n", buffer = bufnr })
+  }, {
+    normal_mode = {
+      buffer = bufnr,
+    },
+  })
 end
 
-function M.common_capabilities()
+function M:common_capabilities()
   local capabilities = vim.lsp.protocol.make_client_capabilities()
   capabilities.textDocument.completion.completionItem.snippetSupport = true
   capabilities.textDocument.completion.completionItem.resolveSupport = {
@@ -74,7 +83,7 @@ function M.common_capabilities()
   return capabilities
 end
 
-function M.get_ls_capabilities(client_id)
+function M:get_ls_capabilities(client_id)
   local client
   if not client_id then
     local buf_clients = vim.lsp.buf_get_clients()
@@ -102,14 +111,14 @@ function M.get_ls_capabilities(client_id)
   return enabled_caps
 end
 
-function M.common_on_init(client, bufnr)
-  if lvim.lsp.on_init_callback then
-    lvim.lsp.on_init_callback(client, bufnr)
+function M:common_on_init(client, bufnr)
+  if self.config.on_init_callback then
+    self.config.on_init_callback(client, bufnr)
     Log:debug "Called lsp.on_init_callback"
     return
   end
 
-  local formatters = lvim.lang[vim.bo.filetype].formatters
+  local formatters = self.config.lang[vim.bo.filetype].formatters
   if not vim.tbl_isempty(formatters) and formatters[1]["exe"] ~= nil and formatters[1].exe ~= "" then
     client.resolved_capabilities.document_formatting = false
     Log:debug(
@@ -118,24 +127,26 @@ function M.common_on_init(client, bufnr)
   end
 end
 
-function M.common_on_attach(client, bufnr)
-  if lvim.lsp.on_attach_callback then
-    lvim.lsp.on_attach_callback(client, bufnr)
+function M:common_on_attach(client, bufnr)
+  if self.config.on_attach_callback then
+    self.config.on_attach_callback(client, bufnr)
     Log:debug "Called lsp.on_init_callback"
   end
-  lsp_highlight_document(client)
+  if self.config.document_highlight then
+    lsp_highlight_document(client)
+  end
   add_lsp_buffer_keybindings(bufnr)
   require("lsp.null-ls").setup(vim.bo.filetype)
 end
 
-function M.configure(lang)
+function M:configure(lang)
   local lsp_utils = require "lsp.utils"
-  local lsp = lvim.lang[lang].lsp
+  local lsp = self.config.lang[lang].lsp
   if (lsp.active ~= nil and not lsp.active) or lsp_utils.is_client_active(lsp.provider) then
     return
   end
 
-  local overrides = lvim.lsp.override
+  local overrides = self.config.override
   if type(overrides) == "table" then
     if vim.tbl_contains(overrides, lang) then
       return
@@ -146,15 +157,18 @@ function M.configure(lang)
     local lspconfig = require "lspconfig"
 
     if not lsp.setup.on_attach then
-      lsp.setup.on_attach = M.common_on_attach
+      lsp.setup.on_attach = function(...)
+        return self:common_on_attach(...)
+      end
     end
     if not lsp.setup.on_init then
-      lsp.setup.on_init = M.common_on_init
+      lsp.setup.on_init = function(...)
+        return self:common_on_init(...)
+      end
     end
     if not lsp.setup.capabilities then
-      lsp.setup.capabilities = M.common_capabilities()
+      lsp.setup.capabilities = self:common_capabilities()
     end
-
     lspconfig[lsp.provider].setup(lsp.setup)
   end
 end
